[AT-3570] Create SAPI method for getting deposit addresses
