+++ lib/binance_api/base.rb
@@ -32,6 +32,56 @@ module BinanceAPI
       params = params.merge(signature: signature)
     end
 
+    def process_request(method, url, params)
+      response = make_request(method, url, params)
+      validate_response(response)
+      build_result response
+    end
+
+    def make_request(method, url, params)
+      signed_params = params_with_signature(params, api_secret)
+      headers = { 'X-MBX-APIKEY' => api_key }
+      safe do
+        if method == :get
+          RestClient.get url, headers.merge(params: signed_params)
+        else
+          RestClient::Request.execute(
+            method: method,
+            url: url,
+            payload: signed_params,
+            headers: headers
+          )
+        end
+      end
+    end
+
+    def validate_response(response)
+      return if response.code == 200
+
+      raise BinanceAPI::RequestError.new(response.code, response.body)
+    end
+
+    def build_result(response)
+      json = JSON.parse(response.body&.empty? ? '{}' : response.body, symbolize_names: true)
+      BinanceAPI::Result.new(json, response.code == 200)
+    end
+
+    def map_params(options)
+      recv_window = options.delete(:recv_window) || BinanceAPI.recv_window
+      timestamp = options.delete(:timestamp) || Time.now
+
+      options.transform_keys! { |k| snake_case_to_camel_case(k) }
+
+      options.merge(recvWindow: recv_window, timestamp: timestamp.to_i * 1000)
+    end
+
+    def snake_case_to_camel_case(key)
+      ['_', '\s'].each do |s|
+        key = key.to_s.gsub(/(?:#{s}+)([a-z])/){ $1.upcase }
+      end
+      key.gsub(/(\A|\s)([A-Z])/){ $1 + $2.downcase }.to_sym
+    end
+
     # ensure to return a response object
     def safe
       yield
