+++ lib/binance_api/version.rb
@@ -1,3 +1,3 @@
 module BinanceAPI
-  VERSION = '1.1.0'.freeze
+  VERSION = '1.1.1'.freeze
 end
