+++ spec/fixtures/vcr_cassettes/BinanceAPISAPI/DepositAddress/WhenInvalidParams/throws_BinanceAPI_RequestError.yml
@@ -1,0 +1,69 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://api.binance.com/sapi/v1/capital/deposit/address?coin=MOJCOIN&recvWindow=5000&signature=d41ee71647026f7136da478ec5d5205409c57eb74d50c18d5ac9db9a4ce90c87&timestamp=1597838191000
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - "*/*"
+      Accept-Encoding:
+      - gzip, deflate
+      User-Agent:
+      - rest-client/2.0.2 (darwin19.5.0 x86_64) ruby/2.6.5p114
+      X-Mbx-Apikey:
+      - b3gxC40YDY5UaGj4uxr5feKb5hM03aJQ8ke79sn3Kbi55IntSvVp3cqXajeTYZpk
+      Host:
+      - api.binance.com
+  response:
+    status:
+      code: 400
+      message: ''
+    headers:
+      Content-Type:
+      - application/json;charset=UTF-8
+      Content-Length:
+      - '50'
+      Connection:
+      - keep-alive
+      Date:
+      - Wed, 19 Aug 2020 11:56:31 GMT
+      Server:
+      - nginx
+      Set-Cookie:
+      - JSESSIONID=7622ECCA7D97B5D2F16D52AC967308D1; Path=/; HttpOnly
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubdomains
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Content-Security-Policy:
+      - default-src 'self'
+      X-Content-Security-Policy:
+      - default-src 'self'
+      X-Webkit-Csp:
+      - default-src 'self'
+      Cache-Control:
+      - no-cache, no-store, must-revalidate
+      Pragma:
+      - no-cache
+      Expires:
+      - '0'
+      X-Cache:
+      - Error from cloudfront
+      Via:
+      - 1.1 c35525724b74ec2ec80741ffbf1ff218.cloudfront.net (CloudFront)
+      X-Amz-Cf-Pop:
+      - WAW50-C1
+      X-Amz-Cf-Id:
+      - WaE1mTVnAa0KR896A6XhckELNFQSTnfMQqhf_UrbRZ3j1mjSqmQemQ==
+    body:
+      encoding: UTF-8
+      string: '{"code":-9000,"msg":"The deposit has been closed"}'
+  recorded_at: Wed, 19 Aug 2020 11:56:31 GMT
+recorded_with: VCR 6.0.0
